# openapi.yaml
openapi: 3.1.0
info:
  version: 0.0.1
  title: art_api
  description: API to manage the art database
  contact:
    name: Ramiro Storni
paths:
...
  '/user/{user_id}':
    get:
      parameters:
        - $ref: '#/components/parameters/path_user_id'
      responses:
        '200':
          $ref: '#/components/responses/get_user_200'
  '/create_user':
    post:
      description: creates a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_user'
      responses:
        '200':
          $ref: '#/components/responses/get_user_200'
  # ARTWORKS
...
  '/artwork/{artwork_id}':
    get:
      parameters:
        - $ref: '#/components/parameters/path_artwork_id'
      responses:
        '200':
          $ref: '#/components/responses/get_artwork_200'
  # AUCTIONS
...
components:
  schemas:
...
    id:
      type: string
      description: UUID for a resource
      example: 123e4567-e89b-12d3-a456-426614174000
    name:
      type: string
      description: The humman readable name or tile associated with an object.
    details:
      type: string
      description: Additional details associated with an object.
    datetime:
      type: string
      format: date-time
      description: Date and time associated with some event.
...
    user:
      type: object
      description: user data model
      properties:
        user_id:
          $ref: '#/components/schemas/id'
        username:
          $ref: '#/components/schemas/name'
        password:
          $ref: '#/components/schemas/name'
        first_name:
          $ref: '#/components/schemas/name'
        last_name:
          $ref: '#/components/schemas/name'
        email:
          $ref: '#/components/schemas/name'
        registration_date:
          $ref: '#/components/schemas/datetime'
        updated_date:
          $ref: '#/components/schemas/datetime'
...
    create_user:
      type: object
      description: data model for creating a user
      properties: 
        username:
          $ref: '#/components/schemas/name'
        password:
          $ref: '#/components/schemas/name'
        first_name:
          $ref: '#/components/schemas/name'
        last_name:
          $ref: '#/components/schemas/name'
        email:
          $ref: '#/components/schemas/name'
      required:
        - username
        - password
        - first_name
        - last_name
        - email
    # Art Object Schemas
    artwork:
      type: object
      description: art object properties
      properties:
        artwork_id:
          $ref: '#/components/schemas/id'
        name:
          $ref: '#/components/schemas/name'
          example: little walter's toys
        details:
          $ref: '#/components/schemas/details'
          example: Pre-WWI still life painting of Aguste Macke's son's toys. 
        artist:
          $ref: '#/components/schemas/name'
          example: Aguste Macke
          default: Unknown
        creation_date:
          $ref: '#/components/schemas/datetime'
          example: 1912-01-01
        medium:
          $ref: '#/components/schemas/details'
          example: oil on canvas
          default: Unknown
        dimentions:
          $ref: '#/components/schemas/details'
          default: Unknown
      required:
        - artwork_id
        - name
...
  parameters:
...
    path_user_id:
      name: user_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/id'
    # Art Object
    path_artwork_id:
      name: artwork_id
      in: path
      required: true
      schema: 
        $ref: '#/components/schemas/id'
    # Auction
...
  responses:
...
    get_user_200:
      description: Successfuly return a user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/user'
    # Art Object
...
    get_artwork_200:
      description: Successfuly returned information about an art object.
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/artwork'
    # Auction
...